+++ components/renku_data_services/connected_services/apispec.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2024-05-30T08:38:44+00:00
+#   timestamp: 2024-06-03T07:31:23+00:00
 
 from __future__ import annotations
 
@@ -9,14 +9,6 @@ from typing import List, Optional
 
 from pydantic import ConfigDict, Field, RootModel
 from renku_data_services.connected_services.apispec_base import BaseAPISpec
-
-
-class RepositoryPermissions(BaseAPISpec):
-    model_config = ConfigDict(
-        extra="forbid",
-    )
-    pull: bool
-    push: bool
 
 
 class ProviderKind(Enum):
@@ -169,45 +161,9 @@ class ConnectedAccount(BaseAPISpec):
     )
 
 
-class RepositoryMetadata(BaseAPISpec):
-    model_config = ConfigDict(
-        extra="forbid",
-    )
-    git_http_url: str = Field(
-        ...,
-        description="A URL which can be opened in a browser, i.e. a web page.",
-        example="https://example.org",
-    )
-    web_url: str = Field(
-        ...,
-        description="A URL which can be opened in a browser, i.e. a web page.",
-        example="https://example.org",
-    )
-    permissions: RepositoryPermissions
-
-
 class ProviderList(RootModel[List[Provider]]):
     root: List[Provider]
 
 
 class ConnectionList(RootModel[List[Connection]]):
     root: List[Connection]
-
-
-class RepositoryProviderMatch(BaseAPISpec):
-    model_config = ConfigDict(
-        extra="forbid",
-    )
-    provider_id: str = Field(
-        ...,
-        description='ID of a OAuth2 provider, e.g. "gitlab.com".',
-        example="some-id",
-    )
-    connection_id: Optional[str] = Field(
-        None,
-        description="ULID identifier",
-        max_length=26,
-        min_length=26,
-        pattern="^[A-Z0-9]{26}$",
-    )
-    repository_metadata: Optional[RepositoryMetadata] = None
