+++ components/renku_data_services/connected_services/blueprints.py
@@ -1,7 +1,7 @@
 """Connected services blueprint."""
 
 from dataclasses import dataclass
-from urllib.parse import unquote, urlparse, urlunparse
+from urllib.parse import urlparse, urlunparse
 
 from sanic import HTTPResponse, Request, json, redirect
 from sanic.log import logger
@@ -9,14 +9,11 @@ from sanic.response import JSONResponse
 from sanic_ext import validate
 
 import renku_data_services.base_models as base_models
-from renku_data_services import errors
 from renku_data_services.base_api.auth import authenticate, only_admins, only_authenticated
 from renku_data_services.base_api.blueprint import BlueprintFactoryResponse, CustomBlueprint
-from renku_data_services.base_api.etag import extract_if_none_match
 from renku_data_services.connected_services import apispec
-from renku_data_services.connected_services.apispec_base import AuthorizeParams, CallbackParams, RepositoryParams
+from renku_data_services.connected_services.apispec_base import AuthorizeParams, CallbackParams
 from renku_data_services.connected_services.db import ConnectedServicesRepository
-from renku_data_services.connected_services.utils import probe_repository
 
 
 @dataclass(kw_only=True)
@@ -184,55 +181,3 @@ class OAuth2ConnectionsBP(CustomBlueprint):
             return json(token.dump_for_api())
 
         return "/oauth2/connections/<connection_id>/token", ["GET"], _get_token
-
-    def get_one_repository(self) -> BlueprintFactoryResponse:
-        """Get the metadata available about a repository."""
-
-        @authenticate(self.internal_gitlab_authenticator)
-        async def _get_internal_gitlab_user(_: Request, user: base_models.APIUser) -> base_models.APIUser:
-            return user
-
-        @authenticate(self.authenticator)
-        @extract_if_none_match
-        async def _get_one_repository(
-            request: Request, user: base_models.APIUser, repository_url: str, etag: str | None
-        ) -> JSONResponse | HTTPResponse:
-            repository_url = unquote(repository_url)
-            RepositoryParams.model_validate(dict(repository_url=repository_url))
-
-            async def get_internal_gitlab_user() -> base_models.APIUser:
-                return await _get_internal_gitlab_user(request)
-
-            result = await self.connected_services_repo.get_repository(
-                repository_url=repository_url, user=user, etag=etag, get_internal_gitlab_user=get_internal_gitlab_user
-            )
-            if result == "304":
-                return HTTPResponse(status=304)
-            headers = (
-                {"ETag": result.repository_metadata.etag}
-                if result.repository_metadata and result.repository_metadata.etag is not None
-                else None
-            )
-            return json(
-                apispec.RepositoryProviderMatch.model_validate(result).model_dump(exclude_none=True, mode="json"),
-                headers=headers,
-            )
-
-        return "/oauth2/api/repositories/<repository_url>", ["GET"], _get_one_repository
-
-    def get_one_repository_probe(self) -> BlueprintFactoryResponse:
-        """Probe a repository to check if it is publicly available."""
-
-        async def _get_one_repository_probe(_: Request, repository_url: str) -> HTTPResponse:
-            repository_url = unquote(repository_url)
-            RepositoryParams.model_validate(dict(repository_url=repository_url))
-
-            result = await probe_repository(repository_url)
-
-            if not result:
-                raise errors.MissingResourceError(
-                    message=f"The repository at {repository_url} does not seem to be publicly accessible."
-                )
-            return HTTPResponse(status=200)
-
-        return "/oauth2/api/repositories/<repository_url>/probe", ["GET"], _get_one_repository_probe
