+++ components/renku_data_services/connected_services/models.py
@@ -79,35 +79,3 @@ class OAuth2TokenSet(dict):
         if self.expires_at is None:
             return None
         return datetime.fromtimestamp(self.expires_at, UTC).isoformat()
-
-
-@dataclass(frozen=True, eq=True, kw_only=True)
-class RepositoryPermissions:
-    """Repository permissions for git operations."""
-
-    pull: bool
-    push: bool
-
-    @classmethod
-    def default(cls) -> "RepositoryPermissions":
-        """Default permissions."""
-        return cls(pull=False, push=False)
-
-
-@dataclass(frozen=True, eq=True, kw_only=True)
-class RepositoryMetadata:
-    """Repository metadata."""
-
-    etag: str | None
-    git_http_url: str
-    web_url: str
-    permissions: RepositoryPermissions
-
-
-@dataclass(frozen=True, eq=True, kw_only=True)
-class RepositoryProviderMatch:
-    """Repository provider match data."""
-
-    provider_id: str
-    connection_id: str | None
-    repository_metadata: RepositoryMetadata | None
