+++ components/renku_data_services/repositories/models.py
@@ -1,0 +1,35 @@
+"""Models for connected services."""
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class RepositoryPermissions:
+    """Repository permissions for git operations."""
+
+    pull: bool
+    push: bool
+
+    @classmethod
+    def default(cls) -> "RepositoryPermissions":
+        """Default permissions."""
+        return cls(pull=False, push=False)
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class RepositoryMetadata:
+    """Repository metadata."""
+
+    etag: str | None
+    git_http_url: str
+    web_url: str
+    permissions: RepositoryPermissions
+
+
+@dataclass(frozen=True, eq=True, kw_only=True)
+class RepositoryProviderMatch:
+    """Repository provider match data."""
+
+    provider_id: str
+    connection_id: str | None
+    repository_metadata: RepositoryMetadata | None
