+++ projects/secrets_storage/pyproject.toml
@@ -23,6 +23,7 @@ packages = [
   { include = "renku_data_services/migrations", from = "../../components" },
   { include = "renku_data_services/namespace", from = "../../components" },
   { include = "renku_data_services/project", from = "../../components" },
+  { include = "renku_data_services/repositories", from = "../../components" },
   { include = "renku_data_services/secrets", from = "../../components" },
   { include = "renku_data_services/session", from = "../../components" },
   { include = "renku_data_services/storage", from = "../../components" },
